Struct beanstalkc::Beanstalkc

pub struct Beanstalkc { /* fields omitted */ }

Beanstalkc provides beanstalkd client operations.

Methods

impl Beanstalkc

pub fn new() -> Beanstalkc

Create a new Beanstalkc instance with default configs. Default connection address is localhost:11300

pub fn host(self, host: &str) -> Self

Change host to beanstalkd server.

Example:

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().host("localhost").connect().unwrap();

pub fn port(self, port: u16) -> Self

Change port to beanstalkd server.

Example:

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().port(12345).connect().unwrap();

pub fn connection_timeout(self, timeout: Option<Duration>) -> Self

Set timeout for TCP connection to beanstalkd server. Default connection timeout is 120s.

Example:

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new()
       .connection_timeout(Some(Duration::from_secs(10)))
       .connect()
       .unwrap();

pub fn connect(self) -> Result<Self, BeanstalkcError>

Connect to a running beanstalkd server.

Examples

Basic usage

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

With custom configurations

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new()
       .host("127.0.0.1")
       .port(11300)
       .connection_timeout(Some(Duration::from_secs(5)))
       .connect()
       .unwrap();

pub fn reconnect(self) -> Result<Self, BeanstalkcError>

Re-connect to the beanstalkd server.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();
let mut conn = conn.reconnect().unwrap();

pub fn put_default(&mut self, body: &[u8]) -> Result<u64, BeanstalkcError>

Put a job into the current tube with default configs. Return job id.

Example:

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let job_id = conn.put_default(b"Rust").unwrap();

pub fn put(
&mut self,
body: &[u8],
priority: u32,
delay: Duration,
ttr: Duration
) -> Result<u64, BeanstalkcError>

Put a job into the current tube and return the job id.

Example

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let job_id = conn.put(
       b"Rust",
       0,
       Duration::from_secs(1),
       Duration::from_secs(10),
   );

pub fn reserve(&mut self) -> Result<Job, BeanstalkcError>

Reserve a job from one of those watched tubes. Return a Job object if it succeeds.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.reserve().unwrap();
// Execute job...
dbg!(job.id());
dbg!(job.body());

job.delete().unwrap();

pub fn reserve_with_timeout(
&mut self,
timeout: Duration
) -> Result<Job, BeanstalkcError>

Reserve a job with given timeout from one of those watched tubes. Return a Job object if it succeeds.

Example

use beanstalkc::Beanstalkc;
use std::time::Duration;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.reserve_with_timeout(Duration::from_secs(10)).unwrap();
// Execute job...
dbg!(job.id());
dbg!(job.body());

job.delete().unwrap();

pub fn kick(&mut self, bound: u32) -> Result<u64, BeanstalkcError>

Kick at most bound jobs into the ready queue.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

assert_eq!(10, conn.kick(10).unwrap());

pub fn kick_job(&mut self, job_id: u64) -> Result<(), BeanstalkcError>

Kick a specific job into the ready queue.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

conn.kick(123).unwrap();

pub fn peek(&mut self, job_id: u64) -> Result<Job, BeanstalkcError>

Return a specific job.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.peek(1).unwrap();
assert_eq!(1, job.id());

pub fn peek_ready(&mut self) -> Result<Job, BeanstalkcError>

Return the next ready job.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.peek_ready().unwrap();
dbg!(job.id());
dbg!(job.body());

pub fn peek_delayed(&mut self) -> Result<Job, BeanstalkcError>

Return the delayed job with the shortest delay left.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.peek_delayed().unwrap();
dbg!(job.id());
dbg!(job.body());

pub fn peek_buried(&mut self) -> Result<Job, BeanstalkcError>

Return the next job in the list of buried jobs.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.peek_buried().unwrap();
dbg!(job.id());
dbg!(job.body());

pub fn do_peek(&mut self, cmd: Command) -> Result<Job, BeanstalkcError>

pub fn tubes(&mut self) -> Result<Vec<String>, BeanstalkcError>

Return a list of all existing tubes.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let tubes = conn.tubes().unwrap();
assert!(tubes.contains(&String::from("default")));

pub fn using(&mut self) -> Result<String, BeanstalkcError>

Return the tube currently being used.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let tube = conn.using().unwrap();
assert_eq!("default".to_string(), tube);

pub fn use_tube(&mut self, name: &str) -> Result<String, BeanstalkcError>

Use a given tube.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let tube = conn.use_tube("jobs").unwrap();
assert_eq!("jobs".to_string(), tube);

pub fn watching(&mut self) -> Result<Vec<String>, BeanstalkcError>

Return a list of tubes currently being watched.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let tubes = conn.watching().unwrap();
assert_eq!(vec!["default".to_string()], tubes);

pub fn watch(&mut self, name: &str) -> Result<u64, BeanstalkcError>

Watch a specific tube.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let watched_count = conn.watch("foo").unwrap();
assert_eq!(2, watched_count);

pub fn ignore(&mut self, name: &str) -> Result<u64, BeanstalkcError>

Stop watching a specific tube.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();
conn.ignore("foo").unwrap();

pub fn stats(&mut self) -> Result<HashMap<String, String>, BeanstalkcError>

Return a dict of statistical information about the beanstalkd server.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

dbg!(conn.stats().unwrap());

pub fn stats_tube(
&mut self,
name: &str
) -> Result<HashMap<String, String>, BeanstalkcError>

Return a dict of statistical information about the specified tube.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

dbg!(conn.stats_tube("default").unwrap());

pub fn pause_tube(
&mut self,
name: &str,
delay: Duration
) -> Result<(), BeanstalkcError>

Pause the specific tube for delay time.

Example

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();
conn.pause_tube("default", Duration::from_secs(100));

pub fn delete(&mut self, job_id: u64) -> Result<(), BeanstalkcError>

Delete job by job id.

Examples

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

conn.delete(123).unwrap();

let mut job = conn.reserve().unwrap();
// Recommended way to delete a job
job.delete().unwrap();

pub fn release_default(&mut self, job_id: u64) -> Result<(), BeanstalkcError>

Release a reserved job back into the ready queue with default priority and delay.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

conn.release_default(1).unwrap();

pub fn release(
&mut self,
job_id: u64,
priority: u32,
delay: Duration
) -> Result<(), BeanstalkcError>

Release a reserved job back into the ready queue.

Example

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

conn.release(1, 0, Duration::from_secs(10)).unwrap();

pub fn bury_default(&mut self, job_id: u64) -> Result<(), BeanstalkcError>

Bury a specific job with default priority.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

conn.bury_default(1).unwrap();

pub fn bury(
&mut self,
job_id: u64,
priority: u32
) -> Result<(), BeanstalkcError>

Bury a specific job.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

conn.bury(1, 0).unwrap();

pub fn touch(&mut self, job_id: u64) -> Result<(), BeanstalkcError>

Touch a job by job_id. Allowing the worker to request more time on a reserved job before it expires.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

conn.touch(1).unwrap();

pub fn stats_job(
&mut self,
job_id: u64
) -> Result<HashMap<String, String>, BeanstalkcError>

Return a dict of statistical information about a job.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let stats = conn.stats_job(1).unwrap();
dbg!(stats);