Struct beanstalkc::Job

pub struct Job<'a> { /* fields omitted */ }

Job is a simple abstraction about beanstalkd job.

Methods

impl<'a> Job<'a>

pub fn new(
conn: &'a mut Beanstalkc,
job_id: u64,
body: String,
reserved: bool
) -> Job

Initialize and return the Job object.

pub fn id(&self) -> u64

Return job id.

pub fn body(&self) -> &str

Return job body.

pub fn reserved(&self) -> bool

Return job reserving status.

pub fn delete(&mut self) -> Result<(), BeanstalkcError>

Delete this job.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.reserve().unwrap();
job.delete().unwrap();

pub fn release_default(&mut self) -> Result<(), BeanstalkcError>

Release this job back to the ready queue with default priority and delay.

Example

use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.reserve().unwrap();
job.release_default().unwrap();

pub fn release(
&mut self,
priority: u32,
delay: Duration
) -> Result<(), BeanstalkcError>

Release this job back to the ready queue with custom priority and delay.

Example

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.reserve().unwrap();
job.release(0, Duration::from_secs(0)).unwrap();

pub fn bury_default(&mut self) -> Result<(), BeanstalkcError>

Bury this job with default priority.

Example

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.reserve().unwrap();
job.bury_default().unwrap();

pub fn bury(&mut self, priority: u32) -> Result<(), BeanstalkcError>

Bury this job with custom priority.

Example

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.reserve().unwrap();
job.bury(1024).unwrap();

pub fn kick(&mut self) -> Result<(), BeanstalkcError>

Kick this job to ready queue.

Example

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.peek_buried().unwrap();
job.kick().unwrap();

pub fn touch(&mut self) -> Result<(), BeanstalkcError>

Touch this reserved job, requesting more time to work on it.

Example

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.reserve().unwrap();
job.touch().unwrap();

pub fn stats(&mut self) -> Result<HashMap<String, String>, BeanstalkcError>

Return a dict of statistical information about this job.

Example

use std::time::Duration;
use beanstalkc::Beanstalkc;

let mut conn = Beanstalkc::new().connect().unwrap();

let mut job = conn.peek_ready().unwrap();
let job_stats = job.stats().unwrap();
dbg!(job_stats);